requirements.txt
motor==2.5.1
PyJWT
bcrypt
uvicorn
pymongo==3.12.3
python-decouple
fastapi
pydantic[email]
passlib
numpy